Put all unit tests in a single binary (Linux) Also put all mocked system calls in a single compilation file. Since all unit tests are in a single binary, they don't need to be in separate files now that unit tests don't selectively choose which mocked calls need to be linked.
