+++ test/files.mk
@@ -1,14 +1,34 @@
+# Define the directories to include and compile
+SRC_DIRS_$(d) := \
+    fly/config \
+    fly/logger \
+    fly/parser \
+    fly/path \
+    fly/socket \
+    fly/system \
+    fly/task \
+    fly/types \
+    test/mock \
+    test/util
+
+# Include the directories containing the unit tests
+SRC_DIRS_$(d) += \
+    test/config \
+    test/logger \
+    test/parser \
+    test/path \
+    test/socket \
+    test/system \
+    test/task \
+    test/traits \
+    test/types
+
 # Define source files
 SRC_$(d) := \
     $(d)/googletest/googletest/src/gtest-all.cc \
     $(d)/googletest/googletest/src/gtest_main.cc
 
-# Define source files for mocked system calls
-ifneq ($(MOCK_SYSTEM_CALLS), )
-
-SRC_$(d) += \
-    $(d)/mock/mock_system.cpp \
-    $(d)/mock/nix/mock_list.cpp \
-    $(foreach mock, $(MOCK_SYSTEM_CALLS), $(d)/mock/nix/mock_$(mock).cpp)
-
-endif
+# Define libraries to link
+LDLIBS_$(d) := \
+    -latomic \
+    -lpthread
