+++ test/system/files.mk
@@ -1,26 +1,2 @@
-# Define the directories to include and compile
-SRC_DIRS_$(d) := \
-    fly/config \
-    fly/logger \
-    fly/parser \
-    fly/path \
-    fly/system \
-    fly/task \
-    fly/types \
-    test/util
-
-# Define linker flags
-LDFLAGS_$(d) += -static-libstdc++
-
-# Define libraries to link
-LDLIBS_$(d) := \
-    -latomic \
-    -lpthread
-
 # Define source files
 $(eval $(call WILDCARD_SOURCES))
-
-# Define mocked system calls
-$(eval $(call MOCK_SYSTEM_CALL, read))
-$(eval $(call MOCK_SYSTEM_CALL, sysinfo))
-$(eval $(call MOCK_SYSTEM_CALL, times))
