+++ test/traits/files.mk
@@ -1,10 +1,2 @@
-# Define the directories to include and compile
-SRC_DIRS_$(d) :=
-
-# Define libraries to link
-LDLIBS_$(d) := \
-    -latomic \
-    -lpthread
-
 # Define source files
 $(eval $(call WILDCARD_SOURCES))
